chore: adds assetRow to dedupe perps and tokens
