+++ app/components/Views/WhatsHappeningDetailView/components/PerpsRow.tsx
@@ -1,25 +1,11 @@
 import React, { useCallback } from 'react';
 import { useNavigation, NavigationProp } from '@react-navigation/native';
-import {
-  AvatarToken,
-  AvatarTokenSize,
-  Box,
-  BoxAlignItems,
-  BoxFlexDirection,
-  BoxJustifyContent,
-  ButtonBase,
-  ButtonBaseSize,
-  FontWeight,
-  Text,
-  TextColor,
-  TextVariant,
-} from '@metamask/design-system-react-native';
 import { PERPS_EVENT_VALUE } from '@metamask/perps-controller';
 import type { RelatedAsset } from '@metamask/ai-controllers';
 import type { PerpsNavigationParamList } from '../../../UI/Perps/types/navigation';
 import Routes from '../../../../constants/navigation/Routes';
 import { strings } from '../../../../../locales/i18n';
-import { getRelatedAssetImageSource } from '../utils/getRelatedAssetImageSource';
+import AssetRow from './AssetRow';
 
 interface PerpsRowProps {
   asset: RelatedAsset;
@@ -33,7 +19,6 @@ interface PerpsRowProps {
  */
 const PerpsRow: React.FC<PerpsRowProps> = ({ asset }) => {
   const navigation = useNavigation<NavigationProp<PerpsNavigationParamList>>();
-
   const hlPerpsMarket = asset.hlPerpsMarket?.[0];
 
   const handleTrade = useCallback(() => {
@@ -47,48 +32,13 @@ const PerpsRow: React.FC<PerpsRowProps> = ({ asset }) => {
     });
   }, [navigation, hlPerpsMarket, asset.name]);
 
-  const rawImageSource = getRelatedAssetImageSource(asset);
-  const imageSource = Array.isArray(rawImageSource)
-    ? (rawImageSource[0] as { uri?: string } | undefined)
-    : (rawImageSource as number | { uri?: string } | undefined);
-
   return (
-    <Box
-      flexDirection={BoxFlexDirection.Row}
-      alignItems={BoxAlignItems.Center}
-      gap={3}
-      twClassName="py-3"
-    >
-      <AvatarToken
-        name={asset.name}
-        size={AvatarTokenSize.Lg}
-        src={imageSource ?? undefined}
-      />
-
-      <Box
-        twClassName="flex-1"
-        flexDirection={BoxFlexDirection.Row}
-        alignItems={BoxAlignItems.Center}
-        justifyContent={BoxJustifyContent.Between}
-      >
-        <Text
-          variant={TextVariant.BodyMd}
-          fontWeight={FontWeight.Medium}
-          color={TextColor.TextDefault}
-        >
-          {asset.symbol}
-        </Text>
-
-        <ButtonBase
-          size={ButtonBaseSize.Md}
-          twClassName="bg-background-default rounded-2xl px-4"
-          onPress={handleTrade}
-          accessibilityLabel={`${strings('bottom_nav.trade')} ${asset.symbol}`}
-        >
-          {strings('bottom_nav.trade')}
-        </ButtonBase>
-      </Box>
-    </Box>
+    <AssetRow
+      asset={asset}
+      actionLabel={strings('bottom_nav.trade')}
+      accessibilityLabel={`${strings('bottom_nav.trade')} ${asset.symbol}`}
+      onAction={handleTrade}
+    />
   );
 };
